🚚[MOVE] : comments.py 분리
