+++ src/app/api/boards.py
@@ -3,7 +3,6 @@ import os
 from fastapi import APIRouter
 from db.sql import select_many_sql, select_one_sql, write_sql
 from schemas.board import BoardCreate, BoardUpdate
-from schemas.comment import CommentCreate, CommentUpdate
 from faker import Faker
 
 router = APIRouter()
@@ -38,27 +37,7 @@ def delete_board(board_id):
     write_sql("DELETE FROM board WHERE id = %s", (board_id))
     return {"msg" : "success"}
 
-@router.get("/comment/{comment_id}")
-def get_comment(comment_id):
-    result = select_one_sql(f"SELECT comment.id, comment.content, board.title, board.content  FROM comment LEFT OUTER JOIN board ON comment.board_id = board.id WHERE comment.id = {comment_id}")
-    return result
-
-
-@router.post("/{board_id}/comment")
-def create_comment(board_id, comment: CommentCreate):
-    write_sql("INSERT INTO comment(content, board_id) VALUES (%s, %s)", (comment.content, board_id))
-    return comment
-
-@router.put("/comment/{comment_id}")
-def update_comment(comment_id, comment: CommentUpdate):
-    write_sql(f"UPDATE comment SET content = %s WHERE id = %s", (comment.content, comment_id))
-    return comment
-
-@router.delete("/comment/{comment_id}")
-def delete_comment(comment_id):
-    write_sql("DELETE FROM comment WHERE id = %s", (comment_id))
-    return {"msg": "success"}
-
+# bulk_board
 @router.post("/bulk/{count_board}")
 def create_boards(count_board : int):
     fake = Faker("ko_KR")
